backend/requirements.txt
pydantic==1.9.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
pyttsx3==2.90
cachetools==4.2.4
pydub
speechrecognition
webrtcvad
transformers
numpy
torch
pyttsx3
cachetools
pyaudio
aiohttp==3.8.1
spacy
gensim
scikit-learn
fastapi==0.68.1
uvicorn==0.15.0
passlib[bcrypt]==1.7.4
pyjwt==2.1.0
sqlalchemy==1.4.23
google-api-python-client
selenium
beautifulsoup4

